renamed the app and all dependencies
